+++ app/frontend/src/javascript/api/payment-schedule.ts
@@ -23,6 +23,11 @@ export default class PaymentScheduleAPI {
     return res?.data;
   }
 
+  static async confirmTransfer (paymentScheduleItemId: number): Promise<CashCheckResponse> {
+    const res: AxiosResponse = await apiClient.post(`/api/payment_schedules/items/${paymentScheduleItemId}/confirm_transfer`);
+    return res?.data;
+  }
+
   static async getItem (paymentScheduleItemId: number): Promise<PaymentScheduleItem> {
     const res: AxiosResponse = await apiClient.get(`/api/payment_schedules/items/${paymentScheduleItemId}`);
     return res?.data;