+++ config/routes.rb
@@ -125,6 +125,7 @@ Rails.application.routes.draw do
       get 'download', on: :member
       get 'items/:id', action: 'show_item', on: :collection
       post 'items/:id/cash_check', action: 'cash_check', on: :collection
+      post 'items/:id/confirm_transfer', action: 'confirm_transfer', on: :collection
       post 'items/:id/refresh_item', action: 'refresh_item', on: :collection
       post 'items/:id/pay_item', action: 'pay_item', on: :collection
     end